Fix x86-64 crash due to RBX clobber by cpuid
Code calling FESTAFILT_Init stored a pointer in %rbx. The top 32 bits of
this pointer were clobbered by the call to cpuid (or rather the macro
wrapping it, which attempted to preserve %ebx by swapping the %ebx
output of CPUID to %esi, neglecting the top half of the register in
64-bit mode). Oops.

So, we're back to __x86_64__ conditionals, unless another approach
presents itself. Sigh.